Cut legacy analyser component
