--- src/components/LegacyAnalyse/index.js
@@ -1,1 +1,0 @@
-export {Analyse} from './Analyse'
