+++ src/components/ReportFlow/ActorList.tsx
@@ -2,7 +2,7 @@ import {Trans} from '@lingui/react'
 import {Message, Segment} from 'akkd'
 import classNames from 'classnames'
 import Color from 'color'
-import JobIcon from 'components/ui/JobIcon'
+import {JobIcon} from 'components/ui/JobIcon'
 import NormalisedMessage from 'components/ui/NormalisedMessage'
 import {Role, RoleKey, ROLES, JobKey, JOBS} from 'data/JOBS'
 import {patchSupported} from 'data/PATCHES'
