+++ src/components/ReportFlow/Analyse/SegmentLinkItem.module.css.d.ts
