+++ src/components/ReportFlow/Analyse/index.ts
@@ -1,0 +1,2 @@
+export {Analyse} from './Analyse'
+export type {AnalyseProps} from './Analyse'
