+++ src/parser/jobs/ast/modules/ArcanaTracking/ArcanaSuggestions.tsx
@@ -1,7 +1,7 @@
 import {t} from '@lingui/macro'
 import {Trans} from '@lingui/react'
 import {ActionLink} from 'components/ui/DbLink'
-import JobIcon from 'components/ui/JobIcon'
+import {JobIcon} from 'components/ui/JobIcon'
 import {Action} from 'data/ACTIONS'
 import {JOBS} from 'data/JOBS'
 import {Analyser} from 'parser/core/Analyser'
